chore: add compressed database to a working branch
I'm discussing this idea with @hellais and I'd like to see
whether it works fine in practice. It might be a good way to
make our building procedure to work for pkg.go.dev and for
the folks who want to build with `go install -v ...` without
cloning the repository first (which is also broken).